pyproject.toml
[project]
name = "fastapi-apiclient"
version = "0.1.0"
description = ""
authors = [
    {name = "thomas gooch",email = "thomas.gooch@mckesson.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "aiobreaker (>=1.2.0,<2.0.0)",
    "fhir-resources (>=8.0.0,<9.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

packages = [
    { include = "app" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"

